editor: Move line export helpers
The editor package now keeps piece-table storage separate from the stateful
Editor facade, while edit.py still owns byte line export helpers and the
public generated-line export function.

Line body conversion and buffer export do not depend on cursor state,
selections, or editor leases. Keeping that byte-export code inside edit.py
makes the stateful editor facade carry stateless output concerns.

This commit addresses that by moving line body conversion, chunk joining,
and export_lines_as_buffer into editor.line_export while letting Editor
reuse those helpers for materialized buffers.

The following tests commit will move export helper coverage to the new
module path.
