+++ src/git_stage_batch/editor/edit.py
@@ -9,6 +9,7 @@ from typing import overload
 from ..core.buffer import LineBuffer
 from ..utils.text import normalize_line_ending
 from .line_endings import detect_line_ending, restore_line_endings_in_chunks
+from .line_export import line_body, line_body_chunks
 from .piece_table import LineLike, LinePieceTable, LineRange, SOURCE_RUN
 
 
@@ -358,7 +359,7 @@ class Editor(Sequence[_LineLike]):
         self._require_open()
         self._require_no_editor_borrowers()
         try:
-            chunks = _line_body_chunks(
+            chunks = line_body_chunks(
                 self._line_bodies(),
                 has_trailing_newline=has_trailing_newline,
                 add_trailing_newline_when_nonempty=(
@@ -467,7 +468,7 @@ class Editor(Sequence[_LineLike]):
 
     def _line_bodies(self) -> Iterator[bytes]:
         for line in self._lines():
-            yield _line_body(line)
+            yield line_body(line)
 
     def _lines(self) -> Iterator[_LineLike]:
         for run_index in range(len(self._pieces)):
@@ -817,29 +818,6 @@ def edit_lines_as_buffer(
         raise
 
 
-def export_lines_as_buffer(
-    lines: Iterable[_LineLike],
-    *,
-    has_trailing_newline: bool = True,
-    add_trailing_newline_when_nonempty: bool = False,
-    line_endings_from: Sequence[bytes] | None = None,
-) -> LineBuffer:
-    """Export generated lines to a buffer without editor state."""
-    chunks = _line_body_chunks(
-        (_line_body(line) for line in lines),
-        has_trailing_newline=has_trailing_newline,
-        add_trailing_newline_when_nonempty=(
-            add_trailing_newline_when_nonempty
-        ),
-    )
-    if line_endings_from is not None:
-        chunks = restore_line_endings_in_chunks(
-            chunks,
-            detect_line_ending(line_endings_from),
-        )
-    return LineBuffer.from_chunks(chunks)
-
-
 def _line_range_bounds(
     start: int | None,
     end: int | None,
@@ -951,7 +929,7 @@ def _spool_inserted_lines(
             if end is not None and index >= end:
                 break
             line_count += 1
-            yield _line_body(line) + b"\n"
+            yield line_body(line) + b"\n"
 
     buffer = LineBuffer.from_chunks(chunks())
     return _InsertedLines(
@@ -985,43 +963,3 @@ def _slice_uses_negative_bounds(line_slice: slice) -> bool:
         (line_slice.start is not None and line_slice.start < 0)
         or (line_slice.stop is not None and line_slice.stop < 0)
     )
-
-
-def _line_body(line: _LineLike) -> bytes:
-    line_bytes = _line_bytes(line)
-    if line_bytes.endswith(b"\r\n"):
-        return line_bytes[:-2]
-    if line_bytes.endswith(b"\n"):
-        return line_bytes[:-1]
-    return line_bytes
-
-
-def _line_bytes(line: _LineLike) -> bytes:
-    if isinstance(line, (bytes, bytearray, memoryview)):
-        return bytes(line)
-    if hasattr(line, "__bytes__"):
-        return bytes(line)
-    raise TypeError(f"expected bytes-compatible line, got {type(line).__name__}")
-
-
-def _line_body_chunks(
-    lines: Iterable[bytes],
-    *,
-    has_trailing_newline: bool,
-    add_trailing_newline_when_nonempty: bool = False,
-) -> Iterator[bytes]:
-    previous_line = b""
-    has_previous_line = False
-    for line in lines:
-        if has_previous_line:
-            yield previous_line + b"\n"
-        previous_line = line
-        has_previous_line = True
-
-    if not has_previous_line:
-        return
-
-    if has_trailing_newline or add_trailing_newline_when_nonempty:
-        yield previous_line + b"\n"
-    else:
-        yield previous_line
